+++ native_library/source/WebViewANE.cpp
@@ -81,6 +81,7 @@ extern "C" {
 			,{ (const uint8_t *) "getCurrentTab", "getCurrentTab", &callSharpFunction }
 			,{ (const uint8_t *) "getTabDetails", "getTabDetails", &callSharpFunction }
 			,{ (const uint8_t *) "shutDown", "shutDown", &callSharpFunction }
+			,{ (const uint8_t *) "clearCache", "clearCache", &callSharpFunction }
 
 		};