+++ CHANGELOG.md
@@ -3,7 +3,7 @@
 - OSX: Deprecate onFullScreen() method, handled internally
 - OSX: Deprecate ON_ESC_KEY event, handled internally
 - AND/iOS/WIN/OSX: added cacheEnabled to settings
-- iOS/OSX: added clearCache() method
+- AND/iOS/OSX: added clearCache() method
 
 ### 0.0.25
 - WIN/OSX: Added multi-tab support